requirements.txt
flask
markdown
google.generativeai
python-dotenv
flask-bcrypt
pymongo
